add checking user_id in controllers

+++ src/services/masterService.js
@@ -1,21 +1,28 @@
 const {master}=require("../model/master")
 class MasterService{
-    static async find_one(masterId){
-        return await master.findOne({master_id:masterId})
-    }
-    static async createMaster(master_id,user_id,name){
 
-        await master.create({
-            master_id:master_id,
-            user_id:user_id,
-            name:name
+    static async createMaster(master_id,user_id,name){
+        return new Promise(async (resolve,reject)=>{
+            let newMaster=new master({
+                master_id:master_id,
+                user_id:user_id,
+                name:name
+            })
+            if(newMaster.save()){
+                resolve(true)
+            }
+            else{
+                reject(false)
+            }   
         })
+
+      
 
     };
 
-    static  deleteMaster(masterId){
+    static  deleteMaster(masterId,userId){
         return new Promise((resolve,reject)=>{
-            master.findOneAndDelete({master_id:masterId},(err,response)=>{
+            master.findOneAndDelete({master_id:masterId,user_id:userId},(err,response)=>{
                 if(response){
                     resolve(response);
                     }
@@ -27,9 +34,9 @@ class MasterService{
        
     };
 
-    static  updateMaster(materId,name){
+    static  updateMaster(materId,userId,name){
         return new Promise((resolve,reject)=>{
-            master.findOneAndUpdate({master_id:materId},{name:name},(err,res)=>{
+            master.findOneAndUpdate({master_id:materId,user_id:userId},{name:name},(err,res)=>{
                 if(res){
                     resolve(res)
                 }
@@ -41,9 +48,9 @@ class MasterService{
       
     };
 
-    static  readMaster(id){
+    static  readMaster(masterId,userId){
         return new Promise((resolve,reject)=>{
-            master.findOne({master_id:id},(err,res)=>{
+            master.findOne({master_id:masterId,user_id:userId},(err,res)=>{
                 if(res){
                     resolve(res)
                 }
